🚧 增加访问日志查询以及数据统计分析

+++ src/Memo.Blog.Application/Anlyanis/Common/SummaryAnlyanisResult.cs
@@ -5,20 +5,20 @@ public record SummaryAnlyanisResult
     /// <summary>
     /// 本周文章数
     /// </summary>
-    public int WeekArticles { get; set; }
+    public long WeekArticles { get; set; }
 
     /// <summary>
     /// 文章数
     /// </summary>
-    public int Articles { get; set; }
+    public long Articles { get; set; }
 
     /// <summary>
     /// 动态总数
     /// </summary>
-    public int Moments { get; set; }
+    public long Moments { get; set; }
 
     /// <summary>
     /// 友链总数
     /// </summary>
-    public int Friends { get; set; }
+    public long Friends { get; set; }
 }
